Re-record tests, add commutativity
