+++ sdk/storage/azure_storage_blob/assets.json
@@ -1,6 +1,6 @@
 {
   "AssetsRepo": "Azure/azure-sdk-assets",
   "AssetsRepoPrefixPath": "rust",
-  "Tag": "rust/azure_storage_blob_49079e88a1",
+  "Tag": "rust/azure_storage_blob_9de297e630",
   "TagPrefix": "rust/azure_storage_blob"
 }
